requirements.txt
# Core ML/DL
tensorflow>=2.8.0
torch>=1.10.0
transformers>=4.15.0
ray[rllib]>=1.9.0
optuna>=2.10.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0
pyarrow>=6.0.0
h5py>=3.6.0
dask>=2021.12.0

# Technical Analysis
ta>=0.7.0
ta-lib>=0.4.24

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.2
plotly>=5.4.0
dash>=2.0.0

# API & Data Collection
python-binance>=1.0.15
ccxt>=1.60.0
requests>=2.26.0
aiohttp>=3.8.0
newsapi-python>=0.2.6
tweepy>=4.12.0
praw>=7.6.0

# Sentiment Analysis
vaderSentiment>=3.3.2
textblob>=0.17.1
nltk>=3.6.0
spacy>=3.4.0
beautifulsoup4>=4.10.0

# Parallel Processing
joblib>=1.1.0
multiprocess>=0.70.12
distributed>=2021.12.0

# Testing & Validation
pytest>=6.2.5
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
hypothesis>=6.30.0
coverage>=7.0.0

# Code Quality & CI/CD
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0
bandit>=1.7.0
safety>=2.3.0
pre-commit>=3.0.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.0.0
mkdocstrings>=0.20.0
mkdocstrings-python>=1.0.0
mike>=1.1.0

# Logging & Monitoring
wandb>=0.12.0
tensorboard>=2.7.0
python-json-logger>=2.0.2

# Utils
tqdm>=4.62.0
python-dotenv>=0.19.0
pyyaml>=6.0

# XAI
shap>=0.40.0
lime>=0.2.0

# Security
cryptography>=41.0.0
certifi>=2023.7.22
urllib3>=2.0.0
